+++ templates/base_kit_package/lib/src/core/providers.dart
@@ -4,6 +4,7 @@ import 'package:provider/single_child_widget.dart';
 
 import '../di/di.dart';
 import 'example/example_cubit.dart';
+import 'logs/logs_state.dart';
 import 'posts/posts_cubit.dart';
 import 'settings/settings_state.dart';
 import 'snack_messages/snack_messages_state.dart';
@@ -15,6 +16,7 @@ abstract class Providers {
       ChangeNotifierProvider<SnackMessagesState>(create: (_) => locator()),
       BlocProvider<ExampleCubit>(create: (_) => locator()),
       BlocProvider<PostsCubit>(create: (_) => locator()),
+      ChangeNotifierProvider<LogsState>(create: (_) => LogsState.instance),
     ];
   }
 }